Edge/requirements.txt
azure-iot-device==2.12.0
python-dotenv==0.20.0
uamqp==1.5.3
azure-eventhub==5.10.0
pandas==1.1.5
openpyxl==3.1.1
paho-mqtt==1.6.1